Adicionar lógica para recuperar detalhes do post via slug

+++ app/Repositories/Admin/PostRepository.php
@@ -60,7 +60,9 @@ class PostRepository
     {
         $post = $this->entity->findOrFail($id);
 
-        Storage::disk('public')->delete($post->featured_image);
+        if($post->featured_image){
+           Storage::disk('public')->delete($post->featured_image);
+        }
 
         $post->delete();
